[WIP] Ast specification
Add support for org-node AST structure validation (is useful for unit tests
and *especially* important for user conversion to the org-node). Right now
I'm workin on the list parsing, but some subnodes are missing - they are
already described in specification, but parser is not ther yet. So, when I
execute `tParseStructure.nim` it shows me all missing subnodes, unexpected
element kinds etc.

```
StmtList
  List
    ListItem items
      Bullet bullet "-"
      Empty counter '' (empty string)
      Paragraph

          wanted {orgEmpty, orgCheckbox} in 2 (checkbox) on path ListItem.[2]
            Optional checkbox

      missing subnode 3 (tag) {orgEmpty, orgParagraph} on path ListItem.[3]
        Described text (for description list)
      missing subnode 4 (header) {orgParagraph} on path ListItem.[4]
        Main part of the list
      missing subnode 5 (completion) {orgEmpty, orgCompletion} on path ListItem.[5]
        Cumulative completion progress for all subnodes
      missing subnode 6 (body) {orgEmpty, orgStmtList} on path ListItem.[6]
        Additional list items - more sublists, or extended body (with
        code blocks, extra parargaphs etc.)
```