differentiate relevant and irrelevant move actions
Tree  diff implementation  supports 'move'  action but  it's relevancy  was
determined by  the ID  of the  parent mapping. If  source node's  parent is
mapped to the  destination node's parent /exactly/ then  they considered to
have the same parents. This rule is what is needed for initial mapping, but
then 'move' also takes into account a /shifted/ position of the node, which
can differ  substantially, resulting  in a verbose  tree mapping  with many
seemingly false triggers.

Now  the relevancy  of the  change  is determined  by the  parent kind  and
position chain instead.